Expose CfnDBSubnetGroup prop for legacy RDS DBs
We have a legacy `RDS Postgres DB` that we want to move under the `CDK`. However, the `CDK` generates its own `CfnDBSubnetGroup` based on the passed `vpcPlacement`. Updates to a subnet require replacing the DB in cloudformation deploys, which we would prefer to avoid doing. We'd like to be able to pass in our existing `CfnDBSubnetGroup` so we can update the db without a replace.